ViewMenu, Program.cs and prio.cs update
Priority.cs had some issue working (PlaceOrder undefined) so i changed it. but please fix it to what its actually supposed to be. Program.cs updated for the CTL and UI. Calling it view menu instead of separating update menu since the use case has it as an extension of View Menu